Made Discrete Collision Detection with selected time on ConservativeAdvancement class public.

+++ src/ccd/conservative_advancement.cpp
@@ -64,7 +64,7 @@ int conservativeAdvancement(const CollisionGeometry* o1,
 
 	toc = continuous_result.getTimeOfContact();
 
-	return result.numContacts();
+	return static_cast<int>(result.numContacts() );
 }
